+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Support for Intelligent Tiering in added to `Object Storage`
 - Support for Relax Parameters for VLAN and make vcnId as optional for LIST VLAN API
 - Support for CloudGuard features release
+- Support for RAC External Database for DB Management
 
 ## 4.23.0 (April 21, 2021)